Allows apps to be filtered to limit metrics
Currently the app sends metrics for all apps in all spaces for a user
and in some cases this might be more than necessary.  This PR allows the
user to specify which apps in which spaces should be used.

The config option app-filter (or APP_FILTER) is optional, but if set
should consist of a comma-separated list of space.appname strings.

e.g.  staging.main_app,staging.test_app